requirements.txt
argparse==1.4.0
numpy==1.17.3
torch==1.3.0
torchvision==0.4.1.post2
Pillow==6.2.1
tqdm==4.37.0
opencv-python==4.1.1.26
screeninfo==0.4.1
glumpy==1.1.0
